resolved GitHubSecret issue
